+++ packages/typespec-ts/src/modular/serialization/build-deserializer-function.ts
@@ -13,9 +13,9 @@ import { useContext } from "../../context-manager.js";
 import { resolveReference } from "../../framework/reference.js";
 import { refkey } from "../../framework/refkey.js";
 import { reportDiagnostic } from "../../lib.js";
-import { NameType, normalizeName } from "../../rlc-common/index.js";
 import { SdkContext } from "../../utils/interfaces.js";
 import { isAzureCoreErrorType } from "../../utils/model-utils.js";
+import { NameType, normalizeName } from "../../utils/name-utils.js";
 import { getAdditionalPropertiesName, normalizeModelName } from "../emit-models.js";
 import {
   getAllAncestors,
@@ -447,7 +447,7 @@ function getAdditionalPropertiesStatement(
   if (typeof deserializerFunction === "string") {
     params.push(deserializerFunction);
   }
-  return context.rlcOptions?.compatibilityMode === true
+  return context.emitterOptions?.compatibilityMode === true
     ? "...item,"
     : `${getAdditionalPropertiesName(context, type)}: ${resolveReference(SerializationHelpers.serializeRecord)}(${params.join(",")}),`;
 }